chore: pass environment as config
